requirements.txt
kafka-python
argparse
# opencv-python